deinline mpoly IO
